+++ src/pages/photo.js
@@ -1,11 +1,25 @@
 import React from 'react'
 import Link from 'gatsby-link'
 import Img from 'gatsby-image'
-import ImagesList from '../components/ImagesList.js'
 
 import Navbar from '../components/navbar.js'
 
 import '../styles/index.css'
+
+function ImagesList(props) {
+  const imgList = props.imgArr.map(
+    img =>
+      img.node.childImageSharp && (
+        <Img
+          key={img.node.childImageSharp.internal.contentDigest}
+          style={{ width: 800, marginBottom: 8 }}
+          sizes={img.node.childImageSharp.sizes}
+          className="photo"
+        />
+      )
+  )
+  return <div>{imgList}</div>
+}
 
 export default ({ data }) => (
   <div>
